feat: build in its own component
