+++ app/components/PlayerActions.jsx
@@ -3,6 +3,7 @@ import fire from '../../fire'
 import PlayerActionsMoveDropUp from './PlayerActionsMoveDropUp'
 import axios from 'axios'
 import { Button, Menu } from 'semantic-ui-react'
+import PlayerActionsBuild from './PlayerActionsBuild'
 
 // TODO: refactor what's on the state to pass down & to actually be efficient and make sense
 // TODO: have buttons activate when available
@@ -62,54 +63,9 @@ export default class PlayerActions extends Component {
     }
   }
 
-  buildResearch = () => {
-    const activePlayer = this.getActivePlayer(this.state.players)
-    const activePlayerCity = activePlayer.position.city
-    const allCities = this.state.cities
-    // does this city already have a research station?
-    const isResearchCity = allCities[activePlayerCity].research === true
-    if (isResearchCity) {
-      return // if so, can't build research station
-    }
-    // do we have the city card to use?
-    const buildInCity = activePlayer.hand.find(function (card) {
-      return card.city === activePlayerCity
-    })
-    if (buildInCity) {
-      // check num research stations, over 6 means we have to reallocate stations
-      let numResearchCenters = 0
-      return fire.database().ref(`/rooms/${this.props.roomName}/state/researchCenters`).once('value', snapshot => {
-        numResearchCenters = snapshot.val()
-      })
-        .then(() => {
-          // set city to have research station
-          return fire.database().ref(`/rooms/${this.props.roomName}/cities/${activePlayerCity}`).update({
-            research: true
-          })
-            .then(() => {
-              // discard used city card by creating newHand without it
-              const newHand = activePlayer.hand.filter(function (card) {
-                return card.city !== buildInCity.city
-              })
-              fire.database().ref(`/rooms/${this.props.roomName}/players/${activePlayer.playerKey}`).update({
-                numActions: activePlayer.numActions - 1,
-                hand: newHand
-              })
-              // add num research stations to game state
-              fire.database().ref(`/rooms/${this.props.roomName}/state`).update({
-                researchCenters: numResearchCenters - 1
-              })
-              // TODO: max num research stations is 6, take away from other cities when over
-            })
-        })
-    } else {
-      // TODO: let player they don't have city card, maybe fade button
-    }
-  }
-
   render() {
     const activePlayer = this.state.players && Object.keys(this.state.players).length && this.getActivePlayer(this.state.players)
-    console.log("RENDERING")
+    const allCities = this.state.cities && this.state.cities
     return (
       <Menu inverted>
         <Menu.Item>
@@ -128,11 +84,7 @@ export default class PlayerActions extends Component {
         </Button>
         </Menu.Item>
         <Menu.Item>
-          <Button
-            onClick={this.buildResearch}
-          >
-            Build
-        </Button>
+          <PlayerActionsBuild allCities={allCities} activePlayer={activePlayer} roomName={this.props.roomName}/>
         </Menu.Item>
         <Menu.Item>
           <Button>
